Login/out to/from the MAS server
Hooked in some logic for logging into and out of the MAS server via the
mas_client subsystem.  The PagesController 'home' action creates a
@mas_client object for the current user (which logs the user into the MAS
server if he is not already logged in) - if the user is signed in.
Information about the client (to be refined) is displayed via the view.
Added logic such that the sign-out process also logs the user out of the
MAS server (MasClient::logout) and deletes the associated MasSession from
the database.

Also, a bug fix: The intraday period-type names in PeriodTypeConstants were
wrong (e.g., 'one-minute' should be: '1-minute').  Fixed by removing the
constant definitions in this module and using the ones (via 'include')
TimePeriodTypeConstants (which resides in library/mas_client).

+++ app/helpers/sessions_helper.rb
@@ -3,13 +3,9 @@ require 'ruby_contracts'
 module SessionsHelper
   include Contracts::DSL
 
-  pre :user_exists do |user| user != nil end
-  #post :signed_in do |user| current_user == user and signed_in? end
-  def sign_in(user)
-    session[:user_id] = user.id
-    @current_user = user
-  end
+  public ###  Access
 
+  # The current, logged-in user
   def current_user
     if @current_user == nil && session != nil
       @current_user = User.find_by_id(session[:user_id])
@@ -17,6 +13,14 @@ module SessionsHelper
     @current_user
   end
 
+  # MasClient object for 'user'
+  def mas_client(user)
+      @mas_client = MasClientTools::mas_client(user: current_user)
+  end
+
+  public ###  Status report
+
+  # Is 'user' the current user?
   def current_user?(user)
     user == current_user
   end
@@ -25,16 +29,26 @@ module SessionsHelper
     session != nil && session[:user_id] != nil
   end
 
-  def signed_in_user
-    unless signed_in?
-      store_location
-      redirect_to signin_path, notice: "Please sign in."
-    end
+  public ###  Basic operations
+
+  # Sign in 'user'
+  pre :user_exists do |user| user != nil end
+  #post :signed_in do |user| current_user == user and signed_in? end
+  def sign_in(user)
+    session[:user_id] = user.id
+    @current_user = user
   end
 
+  # Sign out 'user'
   def sign_out
-    @current_user = nil
-    session.delete(:user_id)
+    if current_user != nil
+      MasClientTools::logout_client(current_user)
+      if current_user.mas_session != nil
+        current_user.mas_session.destroy
+      end
+      @current_user = nil
+      session.delete(:user_id)
+    end
   end
 
   def redirect_back_or_to(default)
